auth by SHA256 with salt

+++ back/lib/Content/sessions.pm
@@ -1,19 +1,46 @@
 ################################################################################
 
-sub select_logon {}
-
-################################################################################
-
 sub do_create_sessions {
 
 	my $data = {
+	
 		success => \1, 
-		timeout => sql_sessions_timeout_in_minutes ()
+		
+		timeout => sql_sessions_timeout_in_minutes (),
+		
+		user    => sql (users => [
+			[login => $_REQUEST {-login}],
+			[fake  => [0, -1]],
+			[LIMIT => 1],
+		]),
+		
 	};
+	
+	if (!$data -> {user} -> {id}) {
+	
+		warn "Non-existing login entered: $_REQUEST{-login}\n";
 
-	$data -> {user} = sql_select_hash ("SELECT * FROM users WHERE login = ? AND password = OLD_PASSWORD(?)", $_REQUEST {-login}, $_REQUEST {-password});
+		return undef;
+	
+	}
+	
+	if ($data -> {user} -> {fake} != 0) {
+	
+		warn "An attempt to use deleted login detected: $_REQUEST{-login}\n";
 
-	$data -> {user} -> {id} or return undef;
+		return undef;
+	
+	}
+
+	my $hash = password_hash ($data -> {user} -> {salt}, $_REQUEST {-password});
+
+	if ($hash ne $data -> {user} -> {password}) {
+	
+		warn "Wrong password entered for $_REQUEST{-login}\n";
+	
+		return undef;
+	
+	}
 
 	sql_do ("DELETE FROM sessions WHERE id_user = ?", $data -> {user} -> {id});
